added support for the ball to become a tool
this is ignorant to the fact that in the minecraft brick breaker we are all playing called life, we-- a figment of the ball, become a tool for the "powerups" (capitalist & patriarchal social infrastructure that exists around us).

We can even see refrence to the powerup methods, often defining our lives by:

OnPickup():
The worlds impact on us as we begin to exist in it.

WhileActive():
Our actions & effects on the world while existing.

OnDeath():
Effect on the world after our death.

How much of your life do you control?